requirements.txt
streamlit>=1.36
pulp>=2.7
pandas>=2.2
numpy>=1.26
xlsxwriter>=3.0.0
reportlab>=4.0.0
openpyxl>=3.1.0
streamlit-authenticator>=0.2.4
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=41.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
